pyproject.toml
[tool.poetry]
name = "dnsbench"
version = "0.1.0"
description = ""
authors = ["Vivek Venugopalan <github@vivekv.info>"]

[tool.poetry.dependencies]
python = "^3.8"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
